chore: further refactor, remove getDefaultPreferences duplication
